ai-news-agent/requirements.txt
feedparser==6.0.10
requests==2.31.0
openai>=1.12.0
python-dateutil==2.8.2
beautifulsoup4==4.12.2
lxml==4.9.3
pytz==2023.3
schedule==1.2.0
flask==3.0.0
flask-cors==4.0.0
